requirements.txt
tensorflow-cpu==2.19.0
numpy==2.1.3
pandas==2.3.0
matplotlib==3.10.3
scikit-learn==1.7.0
tensorboard==2.19.0
streamlit==1.46.0
joblib==1.5.1
scipy==1.16.0